feat: 아군/적군 타워 체력 로직 및 UI 추가

+++ Assets/02. Scripts/Characters/Enemy/Enemy.cs
@@ -16,8 +16,8 @@ public class Enemy : ScriptableObject
     public float HP { get => m_hp; }
 
     [Header("몬스터의 공격력")]
-    [SerializeField] private float m_atk;
-    public float ATK { get => m_atk; }
+    [SerializeField] private int m_atk;
+    public int ATK { get => m_atk; }
 
     [Header("몬스터의 사거리")]
     [SerializeField] private float m_range;
